+++ frontend/test/fixtures.ts
@@ -68,7 +68,7 @@ export function structure(
         entities: {
           active_mode: "sensor.motion_sensor_living_room_active_mode",
           mode_selection: "select.motion_sensor_living_room_mode_selection",
-          automatic: "switch.motion_sensor_living_room_automatic",
+          automatic: "switch.motion_sensor_living_room_follows_preset_mode",
         },
       },
       {
@@ -83,7 +83,7 @@ export function structure(
         entities: {
           active_mode: "sensor.shutter_living_room_active_mode",
           mode_selection: "select.shutter_living_room_mode_selection",
-          automatic: "switch.shutter_living_room_automatic",
+          automatic: "switch.shutter_living_room_follows_preset_mode",
         },
       },
     ],